requirements.txt
pyserial
pymumble_py3
pyaudio